# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List content calendar items
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [idea, scheduled, published, cancelled]
      responses:
        "200":
          description: Content items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
    post:
      operationId: createContent
      tags: [content]
      summary: Create a content calendar item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentBody"
      responses:
        "201":
          description: Created content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
  /content/{id}:
    patch:
      operationId: updateContent
      tags: [content]
      summary: Update a content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentBody"
      responses:
        "200":
          description: Updated content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentItem"
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete a content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /content/upcoming:
    get:
      operationId: getUpcomingContent
      tags: [content]
      summary: Get upcoming scheduled content
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Upcoming content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentItem"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [tool, subscription, hosting, domain, freelancer, ads, other]
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Add an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /focus:
    post:
      operationId: setFocusSlot
      tags: [focus]
      summary: Create or replace one focus slot for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFocusBody"
      responses:
        "200":
          description: Created or updated focus item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFocusItem"
  /focus/{id}:
    patch:
      operationId: updateFocus
      tags: [focus]
      summary: Update a focus item's title or project link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFocusBody"
      responses:
        "200":
          description: Updated focus item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFocusItem"
        "404":
          description: Not found
    delete:
      operationId: deleteFocus
      tags: [focus]
      summary: Clear a focus slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /focus/{id}/toggle:
    post:
      operationId: toggleFocus
      tags: [focus]
      summary: Toggle a focus item's completed state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated focus item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFocusItem"
        "404":
          description: Not found
...
components:
  schemas:
...
    ContentItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        platform:
          type: string
        status:
          type: string
          enum: [idea, scheduled, published, cancelled]
        scheduledDate:
          type: ["string", "null"]
          format: date-time
        projectId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, title, platform, status, createdAt]
    CreateContentBody:
      type: object
      properties:
        title:
          type: string
        platform:
          type: string
        status:
          type: string
          enum: [idea, scheduled, published, cancelled]
          default: idea
        scheduledDate:
          type: ["string", "null"]
          format: date-time
        projectId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
      required: [title, platform]
    UpdateContentBody:
      type: object
      properties:
        title:
          type: string
        platform:
          type: string
        status:
          type: string
          enum: [idea, scheduled, published, cancelled]
        scheduledDate:
          type: ["string", "null"]
          format: date-time
        projectId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [tool, subscription, hosting, domain, freelancer, ads, other]
        billingCycle:
          type: string
          enum: [monthly, yearly, one_time]
        renewalDate:
          type: ["string", "null"]
        active:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, amount, category, billingCycle, active, createdAt, updatedAt]
    CreateExpenseBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [tool, subscription, hosting, domain, freelancer, ads, other]
          default: tool
        billingCycle:
          type: string
          enum: [monthly, yearly, one_time]
          default: monthly
        renewalDate:
          type: ["string", "null"]
        active:
          type: boolean
          default: true
        notes:
          type: ["string", "null"]
      required: [name, amount]
    UpdateExpenseBody:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [tool, subscription, hosting, domain, freelancer, ads, other]
        billingCycle:
          type: string
          enum: [monthly, yearly, one_time]
        renewalDate:
          type: ["string", "null"]
        active:
          type: boolean
        notes:
          type: ["string", "null"]
...
    DailyFocusItem:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        slot:
          type: integer
          enum: [1, 2, 3]
        title:
          type: string
        projectId:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, date, slot, title, createdAt, updatedAt]
...
    SetFocusBody:
      type: object
      properties:
        slot:
          type: integer
          enum: [1, 2, 3]
        title:
          type: string
          minLength: 1
        projectId:
          type: ["integer", "null"]
      required: [slot, title]
    UpdateFocusBody:
      type: object
      properties:
        title:
          type: string
        projectId:
          type: ["integer", "null"]
...